Moved Annotator into Main and incorporate methods that depend on it.
There are two methods in Main that depend on Annotator; Annotator 
doesn't really do anything without these methods. I've therefore 
consolidated them into a single class.
